+++ src/main/java/com/teste/barbearia/service/ClienteService.java
@@ -1,6 +1,8 @@
 package com.teste.barbearia.service;
 
 import java.util.ArrayList;
+
+import com.teste.barbearia.exception.ApiRequestException;
 import java.util.Calendar;
 import java.sql.Date;
 import java.time.LocalDate;
@@ -17,6 +19,7 @@ import java.sql.Time;
 import java.sql.Timestamp;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import org.springframework.web.accept.HeaderContentNegotiationStrategy;
@@ -30,6 +33,7 @@ import com.teste.barbearia.model.entity.Agendamento;
 import com.teste.barbearia.model.entity.Cliente;
 import com.teste.barbearia.model.entity.Prestador;
 import com.teste.barbearia.model.enuns.Horarios;
+import com.teste.barbearia.model.enuns.Mensagens;
 import com.teste.barbearia.model.enuns.NumeroHorariosOfertados;
 
 import br.com.caelum.stella.ValidationMessage;
@@ -49,35 +53,62 @@ public class ClienteService {
   }
   
   public Cliente insere(Cliente cliente) {
-//    TODO: tratar exceçoes
-    CPFValidator cpfValidator = new CPFValidator(); 
-    try {
-      cpfValidator.assertValid(cliente.getCpf()); 
-    }catch (Exception e) {
-      throw e;
+
+    
+    if(!this.clienteDAO.search(cliente.getCpf()).isEmpty()) {
+      throw new ApiRequestException(Mensagens.ERRO_CPF_JA_EXISTE.getMensagem());
     }
+    if(cliente.getCpf().isEmpty() || cliente.getNome().isEmpty())
+      throw new ApiRequestException(Mensagens.ERRO_INPUT_CLIENTE_VAZIO.getMensagem());
+    
+      
+    try {
+      new CPFValidator().assertValid(cliente.getCpf()); 
+    }catch (Exception e) {
+      throw new ApiRequestException(Mensagens.ERRO_CPF_INVALIDO.getMensagem());
+    }
+    
     clienteDAO.save(cliente);
     return clienteDAO.search(cliente.getCpf()).get(0);
   }
   
   public Cliente getCliente(String cpf) throws Exception {
-    // TODO: tratar exceção
+    
+    if(this.clienteDAO.search(cpf).isEmpty())
+      throw new ApiRequestException(Mensagens.ERRO_CPF_NAO_EXISTE.getMensagem());
+    
     Cliente cliente = this.clienteDAO.search(cpf).get(0);
     
-    return clienteDAO.getById(cliente.getId());
+    return clienteDAO.getById(cliente.getId()).get(0);
   }
   public Cliente getCliente(Long id) {
-    // TODO: tratar excecoes
-    return clienteDAO.getById(id);
+    
+    if(this.clienteDAO.getById(id).isEmpty())
+      throw new ApiRequestException(Mensagens.ERRO_ID_NAO_EXISTE.getMensagem());
+    return clienteDAO.getById(id).get(0);
   }
   
   public Cliente updateCliente(Cliente cliente) throws Exception {
-    // TODO: tratar exceçoes
-    CPFValidator cpfValidator = new CPFValidator();
-    List<ValidationMessage> erros = cpfValidator.invalidMessagesFor(cliente.getCpf());
+    
+    if(this.clienteDAO.getById(cliente.getId()).isEmpty()) {
+      throw new ApiRequestException(Mensagens.ERRO_ID_NAO_EXISTE.getMensagem());
+    }
+        
+    try {
+      new CPFValidator().assertValid(cliente.getCpf()); 
+    }catch (Exception e) {
+      throw new ApiRequestException(Mensagens.ERRO_CPF_INVALIDO.getMensagem());
+    }
+    
+    if(!this.clienteDAO.search(cliente.getCpf()).isEmpty()) {
+      // verifica se o cpf que replicou era referente ao antigo registro do usuario no banco
+      if (this.clienteDAO.search(cliente.getCpf()).get(0).getId() != cliente.getId())
+        throw new ApiRequestException(Mensagens.ERRO_CPF_JA_EXISTE.getMensagem());
+      
+    }
     Long id = cliente.getId();
     clienteDAO.update(cliente, id);
-    return this.clienteDAO.getById(id);
+    return this.clienteDAO.getById(id).get(0);
   }
   
   public List<Prestador> listAllPrestadores(){
@@ -86,8 +117,11 @@ public class ClienteService {
   
 
   public String deletarCliente(Long id) throws Exception {
-  // TODO: Tratar exceções
+    
+    if(this.clienteDAO.getById(id).isEmpty())
+      throw new ApiRequestException(Mensagens.ERRO_ID_NAO_EXISTE.getMensagem());
+       
     this.clienteDAO.delete(id);
-   return "Cliente removido com sucesso!";
+   return Mensagens.SUCESSO_REMOVER_CLIENTE.getMensagem();
   }
 }
